django_project/requirements.txt
django >= 3.2, < 3.3
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
pytest==8.3.5
pytest-django==4.11.1
djangorestframework >= 3.12, < 3.13
psycopg2==2.9.10
python-dotenv==1.0.1
factory_boy==3.3.3
flake8==7.3.0
